LayoutTests/http/tests/security/cross-origin-modal-dialog-base-expected.txt

CONSOLE MESSAGE: Window 'showModalDialog' function is deprecated and will be removed soon.
CONSOLE MESSAGE: showModalDialog() is deprecated and will be removed. Please use the <dialog> element instead.
CONSOLE MESSAGE: document.location.origin before setting location to a relative path: http://localhost:8000
CONSOLE MESSAGE: document.location.origin after setting location to a relative path: http://localhost:8000
Tests that a modal dialog doesn't fall back to its openers origin when document.location is set to a relative path.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Showing modal dialog.
Returned from modal dialog.
If test passed, console messages above should both show the same document.location.origin, <http://localhost:8000>.
PASS successfullyParsed is true

TEST COMPLETE
